Keep unactivated users out of ecosystem
Only show activated users in users index and don't allow
non activated users to log in.  Non activated users are redirected
to root after signing up unless signed up through omniauth.

+++ test/integration/users_index_test.rb
@@ -4,6 +4,7 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
   def setup
     @admin = users(:michael)
     @non_admin = users(:archer)
+    @not_activated = users(:lana)
   end
 
   test "index as admin including pagination and delete links" do
@@ -13,10 +14,12 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
     assert_select 'div.pagination'
     first_page_of_users = User.paginate(page: 1)
     first_page_of_users.each do |user|
-      assert_select 'a[href=?]', user_path(user), text: user.email
-      unless user == @admin
-        assert_select 'a[href=?]', user_path(user), text: 'delete',
-                                                    method: :delete
+      unless user == @not_activated 
+        assert_select 'a[href=?]', user_path(user), text: user.email
+        unless user == @admin
+          assert_select 'a[href=?]', user_path(user), text: 'delete',
+                                                      method: :delete
+        end
       end
     end
     assert_difference 'User.count', -1 do
@@ -28,6 +31,14 @@ class UsersIndexTest < ActionDispatch::IntegrationTest
     log_in_as(@non_admin)
     get users_path
     assert_select 'a', text: 'delete', count: 0
+  end
+
+  test "index without activated users" do
+    get signup_path
+    log_in_as(@non_admin)
+    get users_path
+    assert_select 'a', text: "michael@example.com", count: 1
+    assert_select 'a', text: "hands@example.gov", count: 0
   end
 
 end
